Add option to reduce C++ branch coverage problem and add tests

+++ src/versions.cpp
@@ -1,7 +1,7 @@
 #include "PolyLinkRPC/versions.h"
 
 version_number_t versionStringToNumber(const std::string &version) {
-  int major, minor, patch;
+  uint16_t major, minor, patch;
   char dot;
 
   std::istringstream iss(version);
@@ -14,9 +14,9 @@ version_number_t versionStringToNumber(const std::string &version) {
 }
 
 std::string versionNumberToString(version_number_t version) {
-  int major = (version >> 11) & 0x1F;  // Extracting the major part (5 bits)
-  int minor = (version >> 6) & 0x1F;   // Extracting the minor part (5 bits)
-  int patch = version & 0x3F;          // Extracting the patch part (6 bits)
+  uint8_t major = (version >> 11) & 0x1F;  // Extracting the major part (5 bits)
+  uint8_t minor = (version >> 6) & 0x1F;   // Extracting the minor part (5 bits)
+  uint8_t patch = version & 0x3F;          // Extracting the patch part (6 bits)
 
   return std::to_string(major) + "." + std::to_string(minor) + "." +
          std::to_string(patch);
